ImageViewer: refactor
